% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{add_lesson}
\alias{add_lesson}
\title{Add a new lesson}
\usage{
add_lesson(newlesson)
}
\arguments{
\item{newlesson}{A list create by filling up lessonTemplate}
}
\value{

}
\description{
Add a new lesson
}
\examples{
none
}
