% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{addInternalData}
\alias{addInternalData}
\title{Add object to internal package data}
\usage{
addInternalData(dataObj)
}
\arguments{
\item{dataObj}{An object}
}
\value{

}
\description{
Add object to internal package data
}
\examples{
none
}
