% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{addJs2Latest}
\alias{addJs2Latest}
\title{Add JS html to the end of body to the latest modified html}
\usage{
addJs2Latest(jsfile, path = ".")
}
\arguments{
\item{jsfile}{A character. The file path of your html js file}

\item{path}{A character. The path to your latest modified html file}
}
\value{

}
\description{
Add JS html to the end of body to the latest modified html
}
\examples{
none
}
