% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{create_lessonTemplate}
\alias{create_lessonTemplate}
\title{Create a lesson template}
\usage{
create_lessonTemplate()
}
\value{
A list with names
}
\description{
Use this to create a list object with each element named for the detail of a lesson that can be used to fill in so to be used as an input for add_lesson() function
}
\examples{
newlesson <- create_lessonTemplate()
}
