% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webtemplate.R
\name{create_materializedFramework}
\alias{create_materializedFramework}
\title{Create a Materialized Web Framework}
\usage{
create_materializedFramework(exdir = ".")
}
\arguments{
\item{exdir}{A character (default=".") specifies the directory where downloaded files should be placed.}
}
\value{
none. But will download and unzip to recover the framework
}
\description{
Create a Materialized Web Framework
}
\examples{
none
}
