% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{create_snapshot}
\alias{create_snapshot}
\title{Create snapshot from SVG}
\usage{
create_snapshot(svgfilename, width = "100", destfolder = "./snapshot")
}
\arguments{
\item{svgfilename}{A character. file path of the svg file}

\item{width}{A character. default="100" px}

\item{destfolder}{A character. default="./snapshot"}
}
\value{

}
\description{
Create snapshot from SVG
}
\examples{
none
}
