% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medium.R
\name{exec_HTTP}
\alias{exec_HTTP}
\title{Execute HTTP verb based on "VERB resource-path" expression}
\usage{
exec_HTTP(HTTPphrase, query = NULL)
}
\arguments{
\item{HTTPphrase}{A character of "VERB resource-path" expression}

\item{query}{A list of query parameters}
}
\value{
A list
}
\description{
Execute HTTP verb based on "VERB resource-path" expression
}
\examples{
HTTPphrase="GET https://api.medium.com/v1/me"
query <- list(
  accessToken=token
)
exec_HTTP(HTTPphrase, query=query)
}
