% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{get_courseMaterial}
\alias{get_courseMaterial}
\title{Get course material based on list_courses}
\usage{
get_courseMaterial(number = NULL, date = NULL, destfolder = getwd())
}
\arguments{
\item{number}{An integer. represents the row number from list_courses()}

\item{date}{A character in YYYY-MM-DD format represent the date of the course in list_courses()}

\item{destfolder}{A character, path to save the material, default=getwd()}
}
\value{

}
\description{
Get course material based on list_courses
}
\examples{
list_course() -> available_courses
available_courses
get_courseMaterial(number=1)
get_courseMaterial(date="2020-08-26")
get_courseMaterial(nubmer=1, destfolder="/user")

}
