% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_getProjects}
\alias{github_getProjects}
\title{Get all github projects}
\usage{
github_getProjects(github.token)
}
\arguments{
\item{github.token}{A list, returned from auth}
}
\value{

}
\description{
Get all github projects
}
\examples{
none
}
