% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{serveActiveRmdRenderedHTML}
\alias{serveActiveRmdRenderedHTML}
\title{Serve the html generated by the active Rmd in RStudio}
\usage{
serveActiveRmdRenderedHTML(webRoot = NULL)
}
\arguments{
\item{webRoot}{A web root path, default=NULL means the activeRmd dir be root.}
}
\value{

}
\description{
Serve the html generated by the active Rmd in RStudio
}
\examples{
none
}
