% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{serveTheLastModified}
\alias{serveTheLastModified}
\title{serve the last modified html (usually the latest knitted Rmd)}
\usage{
serveTheLastModified(path = ".")
}
\arguments{
\item{path}{A character, default="."}
}
\value{

}
\description{
serve the last modified html (usually the latest knitted Rmd)
}
\examples{
none
}
