% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{webmagic}
\alias{webmagic}
\title{Insert R chunk in Rmd for JS ex post injection and serve}
\usage{
webmagic()
}
\value{

}
\description{
Insert R chunk in Rmd for JS ex post injection and serve
}
\examples{
none
}
