% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webtemplate.R
\name{webtemplateService}
\alias{webtemplateService}
\title{Create a webtemplate service}
\usage{
webtemplateService()
}
\value{

}
\description{
Create a webtemplate service
}
\examples{
wt <- webtemplateService()
}
