% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{serveActiveRmdRenderedHTML}
\alias{serveActiveRmdRenderedHTML}
\title{Serve the html generated by the active Rmd in RStudio}
\usage{
serveActiveRmdRenderedHTML()
}
\value{

}
\description{
Serve the html generated by the active Rmd in RStudio
}
\examples{
none
}
